git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@17628 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/bmp.c
@@ -1115,7 +1115,7 @@ static Image *ReadBMPImage(const ImageInfo *image_info,ExceptionInfo *exception)
               SetPixelIndex(indexes+(x++),index);
               p++;
             }
-          if (x < image->columns)
+          if (x < (ssize_t) image->columns)
             break;
           if (SyncAuthenticPixels(image,exception) == MagickFalse)
             break;
